+++ app/routes/neighborhood.routes.js
@@ -4,6 +4,7 @@ const neighborhoods = require('controllers/neighborhood.controller');
 const router = express.Router();
 
 router.get('/barrios', neighborhoods.findAll);
-router.get('/barrios/:name', neighborhoods.getLngLat);
+router.get('/barrios/:id', neighborhoods.findById);
+// router.get('/barrios/:name', neighborhoods.getLngLat);
 
 module.exports = router;