+++ CHANGELOG.md
@@ -22,6 +22,7 @@ All notable changes to this project are documented in this file.
 - Add endpoint to get department by ID.
 - Add endpoint to get city by ID.
 - Add endpoint to get district by ID.
+- Add endpoint to get neighborhood by ID.
 
 ### Refactored